Updates for 5.6.3, which are mainly changes to support combined Twin/Sophie-Germain prime searches
git-svn-id: https://svn.code.sf.net/p/prpnet/code@146 029e2237-403a-42e8-9450-7768c132b01b

+++ source/PrimeWorkUnitTest.h
@@ -50,7 +50,7 @@ protected:
    uint64_t il_k;
    uint32_t ii_b;
    uint32_t ii_n;
-   int64_t  ii_c;
+   int32_t  ii_c;
    uint32_t ii_d;
 
    bool     ib_HadTestFailure;
