+++ source/TestingProgramFactory.h
@@ -24,7 +24,7 @@ public:
    void     SendPrograms(Socket *theSocket);
 
    void     SetNumber(int32_t serverType, string suffix, string workUnitName,
-                      uint64_t theK, uint32_t theB, uint32_t theN, int64_t theC, uint32_t theD);
+                      uint64_t theK, uint32_t theB, uint32_t theN, int32_t theC, uint32_t theD);
    
    TestingProgram *GetPRPTestingProgram(int32_t serverType,
                                         uint64_t theK, uint32_t theB, uint32_t theN, uint32_t theD);
